requirements_fixed.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
requests==2.31.0
Pillow==10.0.0
pytesseract==0.3.10
easyocr==1.7.0
PyPDF2==3.0.1
pdf2image==1.16.3
numpy<2.0.0
torch>=2.0.0
transformers>=4.30.0
deepspeed>=0.10.0
peft>=0.4.0
datasets>=2.14.0
accelerate>=0.21.0
sentencepiece>=0.1.99
colorama>=0.4.6
tqdm>=4.65.0
pypdfium2>=4.20.0
pymupdf>=1.23.0
python-multipart>=0.0.6
aiofiles>=23.2.1
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pywin32>=306; sys_platform == 'win32'
